Reformat component export.
