+++ src/app/(pages)/contacts/page.tsx
@@ -2,12 +2,10 @@
 
 import EmptyState from "@/components/EmptyState";
 
-const contacts = ()=>{
-    return (
-        <div className=" hidden lg:block lg:pl-80 h-full">
-         <EmptyState/>
-        </div>
-    );
+export default function contacts() {
+  return (
+    <div className=" hidden lg:block lg:pl-80 h-full">
+      <EmptyState />
+    </div>
+  );
 }
-
-export default contacts
